Support codeactions which use workspace edits instead of commands
